방장 권한

+++ ChatoBackend/src/main/java/com/example/ChatoBackend/controller/RoomController.java
@@ -3,15 +3,20 @@ package com.example.ChatoBackend.controller;
 import com.example.ChatoBackend.DTO.MessageDTO;
 import com.example.ChatoBackend.DTO.ParticipantDTO;
 import com.example.ChatoBackend.entity.ChatRoom;
+import com.example.ChatoBackend.entity.User;
+import com.example.ChatoBackend.jwt.JWTUtils;
 import com.example.ChatoBackend.service.ChatRoomServiceImpl;
 import com.example.ChatoBackend.service.MessageServiceImpl;
+import com.example.ChatoBackend.service.UserServiceImpl;
 import com.example.ChatoBackend.store.ConnectedUserAndRoomInfoStore;
+import io.jsonwebtoken.MalformedJwtException;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
@@ -19,6 +24,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.security.sasl.AuthenticationException;
+import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -40,6 +47,10 @@ public class RoomController {
     ChatRoomServiceImpl chatRoomService;
     @Autowired
     MessageServiceImpl messageService;
+    @Autowired
+    UserServiceImpl userService;
+    @Autowired
+    JWTUtils jwtUtils;
 
     @PostMapping("/create")
     public ResponseEntity<Void> createRoom(@Validated @RequestBody ChatRoom chatRoom) throws SQLException {
@@ -55,11 +66,20 @@ public class RoomController {
     }
 
     @PutMapping("/change_subject")
-    public ResponseEntity<Void> changeSubject(@RequestBody Map<String, String> map) {
-        log.info("" + map);
-        chatRoomService.changeSubject(Long.parseLong(map.get("targetRoomId")), map.get("destinationId"));
-        messagingTemplate.convertAndSend("/sub/chat/room/list", map);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Void> changeSubject(
+            @RequestBody Map<String, String> map,
+            HttpServletRequest req) {
+        String token = String.valueOf(req.getHeader(HttpHeaders.AUTHORIZATION));
+        try {
+            long roomId = Long.parseLong(map.get("targetRoomId"));
+            if (!chatRoomService.checkIfIsRoomOwner(roomId, jwtUtils.getUserId(token)))
+                throw new Exception();
+            chatRoomService.changeSubject(roomId, map.get("destinationId"));
+            messagingTemplate.convertAndSend("/sub/chat/room/list", map);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @PostMapping("/enter_password")
@@ -102,14 +122,23 @@ public class RoomController {
     }
 
     @DeleteMapping("/delete/{roomId}")
-    public ResponseEntity<Void> deleteRoom(@PathVariable(ROOM_ID) Long roomId) {
-        chatRoomService.deleteRoom(roomId);
-        messageService.deleteRoom(roomId);
-        Map<String, Integer> map = new HashMap<>();
-        map.put("isDeleted", 1);
-        map.put(ROOM_ID, Math.toIntExact(roomId));
-        messagingTemplate.convertAndSend("/sub/chat/room/list", map);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<Void> deleteRoom(
+            @PathVariable(ROOM_ID) Long roomId,
+            HttpServletRequest req) {
+        String token = String.valueOf(req.getHeader(HttpHeaders.AUTHORIZATION));
+        try {
+            if (!chatRoomService.checkIfIsRoomOwner(roomId, jwtUtils.getUserId(token)))
+                throw new Exception();
+            chatRoomService.deleteRoom(roomId);
+            messageService.deleteRoom(roomId);
+            Map<String, Integer> map = new HashMap<>();
+            map.put("isDeleted", 1);
+            map.put(ROOM_ID, Math.toIntExact(roomId));
+            messagingTemplate.convertAndSend("/sub/chat/room/list", map);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
     }
 
     @GetMapping("/participants/{roomId}")
@@ -122,7 +151,6 @@ public class RoomController {
     public ResponseEntity<byte[]> getContentImage(
             @PathVariable("roomId") String roomId,
             @PathVariable("msgNo") String msgNo) {
-        log.info(roomId + " " + msgNo);
         byte[] imageByteArray = null;
         try {
             String path = "./images/rooms/" + roomId;
